# 64-bit platform optimization for sm2