This pr optimized the performance of SM2 curve with the recommended parameters from GM/T 0003.5-2012, ensuring the security against side-channel attacks while being compatible with most 64-bit architectures and common compilers. Fast modular reduction and modular inverse algorithms specialized for SM2 curve parameters were added to further improve performance. 

Test results showed that with 64-bit platform optimization, SM2 digital signature algorithm had a signing performance improvement of about 241%-286% and a verifying performance improvement of about 74%-97% compared to the baseline version, with peak performance reaching about 23,000 signatures per second and about 12,000 verifications per second in different testing scenarios and tools.

TEST1(speed command):
| Curve  | sign | verify | sign/s | verify/s |
| ------------- | ------------- | ------------- | ------------- | ------------- |
| 256 bits SM2 (Before) | 0.0002s  | 0.0002s | 5966.9  |  6276.8 |
| 256 bits SM2 (After)  | 0.0000s |  0.0001s | 20384.8 | 10946.2 |

TEST2(perf):
| Curve  | sign/s | verify/s |
| ------------- | ------------- | ------------- |
| 256 bits SM2 (Before) |  6017  |  6212 |
| 256 bits SM2 (After)  | 23242 | 12246 |

This PR also fixed the public key expiration bug mentioned in issue [#414](https://github.com/Tongsuo-Project/Tongsuo/issues/435)

##### Checklist
<!-- 基于你的PR的实际情况移除不适用的项目。其他完成的项目修改[ ]为[x]. -->
- [x] 在 https://yuque.com/tsdoc 增加或更新了必要的文档
- [x] 增加或更新了必要的测试用例
- [x] 对于重要修改，更新了CHANGES文件
- [ ] 当前修改存在对已有API参数或返回值的改变
- [x] 当前修改存在对旧版本功能的兼容性改变（如网络协议或密码算法）